Refactor: LectureList 관심사 분리
